Extract restricted editing mode converters and utils.
